chore: fmt + regen bindings

+++ src/test/harnesses/EigenPodHarness.sol
@@ -5,22 +5,19 @@ import "../../contracts/pods/EigenPod.sol";
 import "forge-std/Test.sol";
 
 contract EigenPodHarness is EigenPod {
-
     constructor(
         IETHPOSDeposit _ethPOS,
         IEigenPodManager _eigenPodManager,
         uint64 _GENESIS_TIME
-    ) EigenPod(
-        _ethPOS,
-        _eigenPodManager,
-        _GENESIS_TIME
-    ) {}
+    ) EigenPod(_ethPOS, _eigenPodManager, _GENESIS_TIME) {}
 
     function getActiveValidatorCount() public view returns (uint256) {
         return activeValidatorCount;
     }
 
-    function setActiveValidatorCount(uint _count) public {
+    function setActiveValidatorCount(
+        uint256 _count
+    ) public {
         activeValidatorCount = _count;
     }
 
@@ -30,12 +27,7 @@ contract EigenPodHarness is EigenPod {
         bytes calldata validatorFieldsProof,
         bytes32[] calldata validatorFields
     ) public returns (uint256) {
-        return _verifyWithdrawalCredentials(
-            beaconStateRoot,
-            validatorIndex,
-            validatorFieldsProof,
-            validatorFields
-        );
+        return _verifyWithdrawalCredentials(beaconStateRoot, validatorIndex, validatorFieldsProof, validatorFields);
     }
 
     function setValidatorStatus(bytes32 pkhash, VALIDATOR_STATUS status) public {
